+++ pullrequests/maximum_depth_of_binary_tree/step3.go
@@ -1,0 +1,11 @@
+//lint:file-ignore U1000 Ignore all unused code
+package template
+
+/*
+レビュワーの方へ：
+	- このコードは既にGoの標準のフォーマッタで整形済みです。演算子の周りにスペースがあったりなかったりしますが、これはGoのフォーマッタによるもので、優先順位の高い演算子の周りにはスペースが入らず、低い演算子の周りには入るようになっています。https://qiita.com/tchssk/items/77030b4271cd192d0347
+*/
+
+/*
+COMMENT
+*/
